scripts/requirements.txt
PyGithub>=2.0.0